# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GSNpeeps API
  version: 0.7.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Kontrak HTTP GSNpeeps untuk Employee Database, dashboard HR, absensi,
    approval, notifikasi, dan administrasi akses.

    API menggunakan field JSON `snake_case`, Bearer JWT berumur 8 jam, dan
    cross-check sesi aktif di Redis. Tidak ada refresh-token endpoint.
  contact:
    name: GSNpeeps IT Liaison
servers:
  - url: http://{host}:{port}
    description: Development
    variables:
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://{host}
    description: Production
    variables:
      host:
        default: api.janjikupadamu.id
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/master/lokasi-kantor:
    get:
      operationId: listOfficeLocations
      summary: Ambil master lokasi kantor
      description: |
        Dropdown lokasi WFO untuk seluruh role terautentikasi. Karyawan bebas memilih
        kantor aktif tempat ia hadir; tidak ada assignment kantor permanen. Response dapat
        kosong sampai data lokasi resmi tersedia.
      tags: [Master Organization]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Daftar lokasi kantor aktif dengan koordinat tepercaya.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeLocationListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/absensi/checkin:
    post:
      operationId: recordAttendance
      summary: Catat check-in atau check-out
      description: |
        Karyawan, Atasan, dan HR mencatat absensi milik sendiri. Field `tipe`
        membedakan check-in dan check-out. Hari kerja Senin-Jumat, 09:00-18:00
        Asia/Jakarta. Check-in setelah 09:00 berstatus `terlambat`; checkout sebelum
        18:00 boleh dicatat dengan status `pulang_cepat`. WFO divalidasi maksimal
        100 meter dari kantor aktif yang dipilih. Foto JPG/PNG maksimal 5 MB.
      tags: [Attendance]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AttendanceCheckRequest"
      responses:
        "201":
          description: Absensi berhasil dicatat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/FileTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedFileType"
        "422":
          description: Aturan bisnis absensi gagal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                outOfRadius:
                  value:
                    success: false
                    error:
                      code: OUT_OF_RADIUS
                      message: Lokasi WFO berada di luar radius kantor
                duplicateCheckIn:
                  value:
                    success: false
                    error:
                      code: DUPLICATE_CHECKIN
                      message: Check-in hari ini sudah tercatat
                checkoutWithoutCheckin:
                  value:
                    success: false
                    error:
                      code: CHECKOUT_WITHOUT_CHECKIN
                      message: Check-out memerlukan check-in aktif
                nonWorkingDay:
                  value:
                    success: false
                    error:
                      code: NON_WORKING_DAY
                      message: Absensi reguler hanya tersedia Senin sampai Jumat
                invalidOfficeLocation:
                  value:
                    success: false
                    error:
                      code: INVALID_OFFICE_LOCATION
                      message: Lokasi kantor WFO tidak valid atau tidak aktif
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/akses/role:
    get:
      operationId: listRoles
      summary: Daftar role
      description: HR dan Top Management membaca empat role sistem; Top Management read-only.
      tags: [Access Administration]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Daftar role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/akses/permission:
    get:
      operationId: getPermissionMatrix
      summary: Ambil matriks permission
      description: HR dan Top Management membaca permission per role; Top Management read-only.
      tags: [Access Administration]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Matriks permission.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionMatrixResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      operationId: updatePermission
      summary: Perbarui permission role
      description: Hanya HR. Perubahan permission wajib dicatat pada Audit Log.
      tags: [Access Administration]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionRequest"
      responses:
        "200":
          description: Permission berhasil diperbarui.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdSuccessResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/akses/audit-log:
    get:
      operationId: listAuditLogs
      summary: Daftar audit log
      description: |
        HR dan Top Management membaca audit log yang sudah direduksi; Top Management
        read-only. Audit log append-only dan tidak memiliki mutation API.
      tags: [Access Administration]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
        - name: aksi
          in: query
          schema: { type: string }
        - name: modul
          in: query
          schema: { type: string }
        - name: user_id
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "200":
          description: Daftar audit log.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token berlaku 8 jam. Backend juga memverifikasi session aktif di
        Redis pada setiap request terproteksi.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
...
    DateFromParam:
      name: tanggal_mulai
      in: query
      schema: { type: string, format: date }
    DateToParam:
      name: tanggal_selesai
      in: query
      schema: { type: string, format: date }
...
  responses:
    BadRequest:
      description: Request atau parameter tidak valid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            invalidRequest:
              value:
                success: false
                error: { code: INVALID_REQUEST, message: Request tidak valid }
    Unauthorized:
      description: Kredensial tidak ada, tidak valid, atau kedaluwarsa.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unauthorized:
              value:
                success: false
                error: { code: UNAUTHORIZED, message: Autentikasi diperlukan }
    Forbidden:
      description: User tidak memiliki kewenangan atau akses row-level.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            forbidden:
              value:
                success: false
                error: { code: FORBIDDEN, message: Akses ditolak }
    NotFound:
      description: Resource tidak ditemukan.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            notFound:
              value:
                success: false
                error: { code: NOT_FOUND, message: Data tidak ditemukan }
...
    ValidationError:
      description: Validasi bisnis atau field gagal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validation:
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Validasi gagal
                  fields: { tanggal_mulai: Tanggal mulai wajib diisi }
    FileTooLarge:
      description: File melebihi batas 5 MB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            fileTooLarge:
              value:
                success: false
                error: { code: FILE_TOO_LARGE, message: Ukuran file maksimal 5 MB }
    UnsupportedFileType:
      description: Tipe file tidak didukung.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            unsupported:
              value:
                success: false
                error: { code: UNSUPPORTED_FILE_TYPE, message: Tipe file tidak didukung }
    TooManyRequests:
      description: Rate limit terlampaui atau akun terkunci sementara.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            rateLimited:
              value:
                success: false
                error: { code: RATE_LIMITED, message: Terlalu banyak request }
    InternalError:
      description: Kesalahan internal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            internal:
              value:
                success: false
                error: { code: INTERNAL_ERROR, message: Terjadi kesalahan internal }
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 0193f406-8d2f-7d0c-bf34-61ff192e46de
...
    RoleName:
      type: string
      enum: [karyawan, atasan, hr, top_management]
...
    WorkMode:
      type: string
      enum: [WFO, WFH, WFA]
    AttendanceType:
      type: string
      enum: [check_in, check_out]
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Validasi gagal }
        fields:
          $ref: "#/components/schemas/ValidationFields"
    ValidationFields:
      type: object
      additionalProperties: { type: string }
      example: { email: Format email tidak valid }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error: { $ref: "#/components/schemas/ErrorObject" }
    PaginationMeta:
      type: object
      required: [page, limit, total_data, total_page]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total_data: { type: integer, minimum: 0 }
        total_page: { type: integer, minimum: 0 }
    IdData:
      type: object
      required: [id]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
    IdSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/IdData" }
...
    OfficeLocation:
      type: object
      required: [id, kode, nama, latitude, longitude, is_active]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        kode: { type: string, maxLength: 50 }
        nama: { type: string, maxLength: 150 }
        alamat:
          type: [string, "null"]
          description: Dapat dilengkapi setelah data alamat kantor resmi tersedia.
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        is_active: { type: boolean }
...
    OfficeLocationListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/OfficeLocation" } }
...
    Attendance:
      type: object
      required: [id, employee_id, tanggal, tipe, mode_kerja, waktu, status]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        employee_id: { $ref: "#/components/schemas/UUID" }
        tanggal: { type: string, format: date }
        tipe: { $ref: "#/components/schemas/AttendanceType" }
        mode_kerja: { $ref: "#/components/schemas/WorkMode" }
        waktu: { type: string, format: date-time }
        gps_lat: { type: number, minimum: -90, maximum: 90 }
        gps_long: { type: number, minimum: -180, maximum: 180 }
        office_location_id:
          oneOf: [{ $ref: "#/components/schemas/UUID" }, { type: "null" }]
          description: Lokasi kantor yang dipilih untuk WFO; null untuk WFH/WFA.
        distance_meters:
          type: [number, "null"]
          minimum: 0
          description: Jarak hasil hitung server untuk WFO; null untuk WFH/WFA.
        foto_url: { type: string, format: uri }
        status: { type: string, enum: [tepat_waktu, terlambat, pulang_cepat, valid] }
    AttendanceCheckRequest:
      type: object
      required: [tipe, mode_kerja, gps_lat, gps_long, foto]
      description: |
        Koordinat wajib untuk seluruh mode. Backend menerapkan radius kantor maksimal
        100 meter hanya ketika mode_kerja=WFO terhadap lokasi kantor aktif yang dipilih;
        WFH dan WFA tetap menyimpan koordinat tetapi tidak menjalankan validasi radius.
      properties:
        tipe: { $ref: "#/components/schemas/AttendanceType" }
        mode_kerja: { $ref: "#/components/schemas/WorkMode" }
        gps_lat: { type: number, minimum: -90, maximum: 90 }
        gps_long: { type: number, minimum: -180, maximum: 180 }
        office_location_id:
          $ref: "#/components/schemas/UUID"
          description: Wajib untuk WFO dan tidak digunakan untuk WFH/WFA.
        foto: { type: string, format: binary, description: JPG/PNG maksimal 5 MB. }
      allOf:
        - if:
            properties:
              mode_kerja: { const: WFO }
            required: [mode_kerja]
          then:
            required: [office_location_id]
    AttendanceCheckResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Attendance" }
...
    RoleSummary:
      type: object
      required: [id, nama, deskripsi]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        nama: { $ref: "#/components/schemas/RoleName" }
        deskripsi: { type: string }
    RoleListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/RoleSummary" } }
    Permission:
      type: object
      required: [id, role_id, modul, aksi, is_allowed]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        role_id: { $ref: "#/components/schemas/UUID" }
        modul: { type: string }
        aksi: { type: string, enum: [create, read, update, delete, approve, export] }
        is_allowed: { type: boolean }
    PermissionMatrixResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Permission" } }
    UpdatePermissionRequest:
      type: object
      additionalProperties: false
      required: [role_id, modul, aksi, is_allowed]
      properties:
        role_id: { $ref: "#/components/schemas/UUID" }
        modul: { type: string, maxLength: 100 }
        aksi: { type: string, enum: [create, read, update, delete, approve, export] }
        is_allowed: { type: boolean }
    AuditLog:
      type: object
      required: [id, user_id, aksi, modul, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        # Aktor sistem (auto-escalation, scheduler) tidak memiliki user; lihat keputusan D-029.
        user_id: { oneOf: [{ $ref: "#/components/schemas/UUID" }, { type: "null" }] }
        nama_user: { type: [string, "null"] }
        aksi: { type: string }
        modul: { type: string }
        resource_id: { oneOf: [{ $ref: "#/components/schemas/UUID" }, { type: "null" }] }
        detail: { type: [object, "null"], additionalProperties: true }
        ip_address: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
    AuditLogListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessResponse"
        - type: object
          required: [meta]
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/AuditLog" } }
            meta: { $ref: "#/components/schemas/PaginationMeta" }
...